+++ source/idstorage.hpp
@@ -3,11 +3,11 @@
 #include <nds.h>
 class IdStorage
 {
-    u8 N = 127;
+    int N = 127;
     bool ids[127] = {false};
 
 public:
     s16 getId();
 
-    void releaseId(u8 i);
+    void releaseId(int i);
 };
